requirements.txt
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
attrs==24.3.0
cattrs==24.1.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
dnspython==2.4.2
fastapi==0.103.2
h11==0.14.0
httptools==0.6.0
idna==3.4
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
motor==3.3.2
ochrona==2.0.2
packaging==24.2
platformdirs==4.3.6
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.18.0
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
requests-cache==1.2.1
rich==13.9.4
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tarsafe==0.0.5
toml==0.10.2
typing_extensions==4.8.0
url-normalize==1.4.3
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
websockets==11.0.3
